Implemented conversion between hex and decimal numbers
This was the first crude attempt using standard techniques for converting between different number bases.

It was no great surprise converting the decimal number to the hex number was slower than using `ToString` with a format string, as well as consuming more memory.

It was a little surprising that converting a hex number to a decimal was faster than using `int.Parse` with `NumberStyles`. However, I haven't yet got any validation in place which will slow things down.